+++ messages/de/hotspot.php
@@ -9,6 +9,7 @@ return [
     'COMMON_HOTSPOT' => 'Hotspot',
     'COMMON_HOTSPOT_ASSET' => 'Hotspot Asset',
     'HOTSPOT_DELETE_TITLE' => 'Hotspot löschen',
+    'HOTSPOT_SUCCESS_CREATED' => 'Der Hotspot wurde erstellt.',
     'HOTSPOT_SUCCESS_DELETED' => 'Der Hotspot wurde gelöscht.',
     'HOTSPOT_SUCCESS_DUPLICATED' => 'Der Hotspot wurde dupliziert.',
     'HOTSPOT_SUCCESS_UPDATED' => 'Der Hotspot wurde aktualisiert.',