+++ messages/en-US/hotspot.php
@@ -9,6 +9,7 @@ return [
     'COMMON_HOTSPOT' => 'Hotspot',
     'COMMON_HOTSPOT_ASSET' => 'Hotspot Asset',
     'HOTSPOT_DELETE_TITLE' => 'Delete Hotspot',
+    'HOTSPOT_SUCCESS_CREATED' => 'The hotspot was created.',
     'HOTSPOT_SUCCESS_DELETED' => 'The hotspot was deleted.',
     'HOTSPOT_SUCCESS_DUPLICATED' => 'The hotspot was duplicated.',
     'HOTSPOT_SUCCESS_UPDATED' => 'The hotspot was updated.',